+++ Makefile
@@ -3,16 +3,16 @@ install:
 		pip install -r requirements.txt
 
 test:
-	python -m pytest -vv --cov=main --cov=mylib test_*.py
+	python -m pytest -vv --cov=main --cov=src test_*.py
 
 format:	
 	black *.py 
 
 lint:
 	#disable comment to test speed
-	#pylint --disable=R,C --ignore-patterns=test_.*?py *.py mylib/*.py
+	#pylint --disable=R,C --ignore-patterns=test_.*?py *.py src/*.py
 	#ruff linting is 10-100X faster than pylint
-	ruff check *.py mylib/*.py
+	ruff check *.py src/*.py
 
 container-lint:
 	docker run --rm -i hadolint/hadolint < Dockerfile
